ai/voice-server/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
pyyaml
soundfile
tensorflow
tensorflowtts
pororo
pydub
